update stock count amount when try to release order, now the counter does not include the order current item amount

+++ src/main/java/service/OrderItemService.java
@@ -71,7 +71,7 @@ public class OrderItemService extends ApplicationService<OrderItem, OrderItemDAO
 		update(orderItem);
 	}
 	
-	public boolean isValidForRelease(OrderItem orderItem, BigDecimal itemAmount) {
+	public boolean isValidForRelease(OrderItem orderItem, Long itemAmount) {
 		return orderItem.getAmount() <= itemAmount.longValue();	
 	}
